Function modular_arithmetic::mod_inverse

pub fn mod_inverse(a: u64, m: u64) -> Option<u64>

Calculates the inverse of a with respect to modulus m, if it exists.

Uses an adaptation of the extended Euclidean algorithm, modified to avoid using signed integers.

Examples

use modular_arithmetic::mod_inverse;

assert_eq!(mod_inverse(1, 7), Some(1));
assert_eq!(mod_inverse(2, 7), Some(4));
assert_eq!(mod_inverse(3, 7), Some(5));
assert_eq!(mod_inverse(4, 7), Some(2));
assert_eq!(mod_inverse(5, 7), Some(3));
assert_eq!(mod_inverse(6, 7), Some(6));